requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.3.3
Flask-DotEnv==0.1.2
Flask-SocketIO==5.3.6
frozenlist==1.3.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.0rc3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.28.0
psycopg2==2.9.7
Pygments==2.16.1
pymdown-extensions==10.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-geoip-python3==1.3
python-socketio==5.8.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.30.0
six==1.16.0
soupsieve==2.5
tiktoken==0.5.1
tqdm==4.65.0
urllib3==2.0.2
uWSGI==2.0.22
websocket-client==1.5.2
Werkzeug==2.3.7
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
